+++ src/ContractSaveFix/Plugin.cs
@@ -1,6 +1,7 @@
 ﻿using System.Reflection;
 using BepInEx;
 using BepInEx.Logging;
+using DungeonGeneration;
 using HarmonyLib;
 using YAPYAP;
 
@@ -10,12 +11,23 @@ namespace ContractSaveFix
     internal partial class Plugin : BaseUnityPlugin
     {
         internal static ManualLogSource Log { get; private set; }
+        private static bool ServerStarted = false;
 
         private void Awake()
         {
             Log = Logger;
             Harmony.CreateAndPatchAll(Assembly.GetExecutingAssembly());
-            Log.LogInfo($"Plugin {Name} is loaded, yay!");
+            Log.LogInfo($"Plugin {Name} is loaded, fixing your contracts!");
+        }
+
+        [HarmonyPatch(typeof(DungeonTasks), nameof(DungeonTasks.Awake))]
+        internal class DungeonTasksAwakeHook
+        {
+            public static void Prefix(DungeonTasks __instance)
+            {
+                Log.LogDebug($"{__instance} Awake");
+                ServerStarted = false;
+            }
         }
 
         [HarmonyPatch(typeof(DungeonTasks), nameof(DungeonTasks.OnStartServer))]
@@ -24,38 +36,25 @@ namespace ContractSaveFix
             public static void Prefix(DungeonTasks __instance)
             {
                 Log.LogDebug($"{__instance} OnStartServer");
-
-                Log.LogInfo("Creating delayed instance information");
-                if (DungeonTasks.Instance == null)
-                {
-                    DungeonTasks.Instance = __instance;
-                    __instance.allTasks = [.. __instance.constantTasks, .. __instance.randomTasks, .. __instance.collectableTasks];
-                    for (int i = 0; i < __instance.allTasks.Length; i++)
-                    {
-                        GameplayTaskSO gameplayTaskSO = __instance.allTasks[i];
-                        if (gameplayTaskSO != null)
-                        {
-                            gameplayTaskSO.TaskId = i;
-                        }
-                    }
-                }
-                else
-                {
-                    UnityEngine.Object.Destroy(__instance.gameObject);
-                }
-
-                Log.LogMessage("Creating dungeon tasks at session start");
-                __instance.CreateTasks(DungeonManager.Instance.Generator);
+                ServerStarted = true;
+                Log.LogMessage($"Creating tasks now!");
+                __instance.CreateTasks(DungeonManager.Instance.generator);
             }
         }
 
-        [HarmonyPatch(typeof(DungeonTasks), nameof(DungeonTasks.Awake))]
-        internal class DungeonTasksAwakeDelete
+        [HarmonyPatch(typeof(DungeonTasks), nameof(DungeonTasks.CreateTasks))]
+        internal class DungeonTasksCreateTasksHook
         {
-            public static bool Prefix(DungeonTasks __instance)
+            public static bool Prefix(DungeonTasks __instance, DungeonGenerator generator, bool forceNewGeneration = false)
             {
-                Log.LogInfo($"{__instance}, skipping Awake");
-                return false;
+                Log.LogDebug($"{__instance} CreateTasks");
+                if(!ServerStarted)
+                {
+                    Log.LogWarning("Skipping CreateTasks that is too early!");
+                    return false;
+                }
+
+                return true;
             }
         }
     }